+++ spec/acceptance/support/pkg_rpm_helpers.rb
@@ -1,5 +1,8 @@
 module Simp::BeakerHelpers::SimpRakeHelpers::PkgRpmHelpers
 
+  require_relative 'build_user_helpers'
+  include Simp::BeakerHelpers::SimpRakeHelpers::BuildUserHelpers
+
   # rake command string to run on hosts
   # passes on useful troubleshooting env vars
   def rake_cmd
@@ -14,17 +17,6 @@ module Simp::BeakerHelpers::SimpRakeHelpers::PkgRpmHelpers
       end
     end
     cmd
-  end
-
-
-  def copy_host_files_into_build_user_homedir(
-    hosts,
-    root_dir=File.expand_path('../../../',__FILE__)
-  )
-    # I've added the `ch* -R` on the SUT-side, which seems to work on a fresh checkout
-    on hosts, 'cp -a /host_files /home/build_user/; ' +
-             'chmod -R go=u-w /home/build_user/host_files/{dist,**/dist,junit,log}; ' +
-             'chown -R build_user:build_user /home/build_user/host_files; '
   end